add prob4: 의상
